Feature: For loops in Python3 module (#16)
This is kinda a giant biggie biggie huge milestone. For loop!!

The biggest ground was laid by last PR (#15) which setup most code for enumerables/iterators. This PR just added compilation of for-loops. And a bunch of refactoring.

Example of code that currently works:

```py
sum = 0
for c in str(int(0xC0FFEE)):
    sum = sum + int(c)

# sum is 28
``` 

# Python3 module
- Added compilation of `forStmt` syntax rules
- Added compilation of `exprList` syntax rules
- Added compilation of `testList` syntax rules
- Added support for `None` literals
- Cleaned up statement compilation tests
- Cleaned up some syntax rule parsing.
- Upped version to `0.4.1`